backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
Flask-JWT-Extended==4.7.1
APScheduler==3.11.0
Werkzeug==3.1.3
Pillow==10.1.0
easyocr==1.7.1
PyMuPDF==1.24.2
google-generativeai
youtube-transcript-api
gunicorn
python-dotenv
psycopg2-binary
